refactor(shipping): Phase 8 relocate InternalOutboxEndpoints + Program.cs manifest
- Move Endpoints/InternalOutboxEndpoints.cs to Infrastructure/Outbox/
  with namespace Shipping.Service.Infrastructure.Outbox. Delete now-empty
  Endpoints/ folder. RequireService policy gate preserved verbatim.
- Reshape Program.cs into a manifest: all 15 AddXxxSlice() registrations
  chained into one fluent block after platform infra (DB, outbox, event
  bus, observability, health, auth, carrier singletons, hosted services),
  followed by Map* mapping block, RegisterInternalOutboxEndpoints,
  HTTPS/JWT. Behavior unchanged.
- Build green; 107/107 Shipping.Tests pass (7 Phase-10 skips unchanged).
